+++ oci/database_vm_cluster_resource.go
@@ -122,6 +122,10 @@ func DatabaseVmClusterResource() *schema.Resource {
 				Type:     schema.TypeInt,
 				Computed: true,
 			},
+			"last_patch_history_entry_id": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
 			"lifecycle_details": {
 				Type:     schema.TypeString,
 				Computed: true,
@@ -473,6 +477,10 @@ func (s *DatabaseVmClusterResourceCrud) SetData() error {
 
 	if s.Res.IsSparseDiskgroupEnabled != nil {
 		s.D.Set("is_sparse_diskgroup_enabled", *s.Res.IsSparseDiskgroupEnabled)
+	}
+
+	if s.Res.LastPatchHistoryEntryId != nil {
+		s.D.Set("last_patch_history_entry_id", *s.Res.LastPatchHistoryEntryId)
 	}
 
 	s.D.Set("license_model", s.Res.LicenseModel)